Enable phpstan

+++ src/VerificationCodeManager.php
@@ -9,23 +9,36 @@ use Zing\LaravelSms\Notifications\VerificationCode;
 
 class VerificationCodeManager
 {
+    /**
+     * @var \Illuminate\Contracts\Cache\Repository
+     */
     protected $cacheManager;
 
     /**
      * VerificationCodeManager constructor.
      *
-     * @param $cacheManager
+     * @param \Illuminate\Contracts\Cache\Repository $cacheManager
      */
     public function __construct(Repository $cacheManager)
     {
         $this->cacheManager = $cacheManager;
     }
 
+    /**
+     * @param \Overtrue\EasySms\Contracts\PhoneNumberInterface|string $number
+     * @return string
+     */
     protected function getPrefixedKey($number): string
     {
         return config('sms.verification.prefix') . $number;
     }
 
+    /**
+     * @param \Overtrue\EasySms\Contracts\PhoneNumberInterface|string $number
+     * @param int|string $code
+     * @throws \Psr\SimpleCache\InvalidArgumentException
+     * @return bool
+     */
     public function verify($number, $code): bool
     {
         if (config('sms.verification.debug', false)) {
@@ -40,12 +53,18 @@ class VerificationCodeManager
         return (int) $code === (int) $issuedCode;
     }
 
+    /**
+     * @param \Overtrue\EasySms\Contracts\PhoneNumberInterface|string|int $number
+     * @param int|null $ttl
+     * @throws \Psr\SimpleCache\InvalidArgumentException
+     * @return int
+     */
     public function issue($number, $ttl = null): int
     {
         $length = config('sms.verification.length');
         $code = random_int(10 ** ($length - 1), (10 ** $length) - 1);
         if ($ttl === null) {
-            $ttl = config('sms.verification.ttl');
+            $ttl = (int) config('sms.verification.ttl');
         }
 
         $this->cacheManager->set($this->getPrefixedKey($number), $code, $ttl * 60);
